// types/interfaces.ts
interface ProfileInfo {
  username: string;
  email: string;
  address: string;
  role: string;
  isAuthorVerified: boolean;
  isVerified: boolean;
  _id: string;
  createdAt: string;
  updatedAt: string;
}